@simon-mo 2024-05-30
It needs to be opt-in (i think it is already?), similar to TGI:
> text-generation-inference is instrumented with distributed tracing using OpenTelemetry. You can use this feature by setting the address to an OTLP collector with the --otlp-endpoint argument.