@ronensc 2024-06-11
Since previous status:
1. Decoupled tracing from `log_stats`.
1. Added support for grpc exporter.
1. Added a one-time warning log for requests with a tracing header when tracing is disabled.
1. Updated the markdown guide.
1. Added a simple e2e unit test.
1. Measured the e2e overhead of sending traces:

Baseline: Without `--otlp-traces-endpoint`
```
$ OTEL_EXPORTER_OTLP_TRACES_INSECURE=true python3 benchmarks/benchmark_latency.py \
	--input-len 512 \
	--output-len 128 2>&1

...
Avg latency: 0.5273534125333412 seconds
10% percentile latency: 0.5149867414999904 seconds
25% percentile latency: 0.5197465127499754 seconds
50% percentile latency: 0.5262749200000485 seconds
75% percentile latency: 0.5300818937500082 seconds
90% percentile latency: 0.5346117969000375 seconds
```

With `--otlp-traces-endpoint`
```
$ OTEL_EXPORTER_OTLP_TRACES_INSECURE=true python3 benchmarks/benchmark_latency.py \
	--input-len 512 \
	--output-len 128 \
	--otlp-traces-endpoint grpc://$JAEGER_IP:4317 2>&1

...
Avg latency: 0.5384878202333273 seconds
10% percentile latency: 0.5267181372999744 seconds
25% percentile latency: 0.5291085377500337 seconds
50% percentile latency: 0.5376283054999362 seconds
75% percentile latency: 0.5437845339999967 seconds
90% percentile latency: 0.5500757966999685 seconds
```


I ran these experiments on my laptop using WSL2, with Jaeger running in Docker.
GPU: NVIDIA RTX A4000