@ronensc 2024-06-12
It appears that the CI is failing due to an issue with tracing. I'll investigate and add a fix.