@AndSonder 2024-11-17
@rkooo567  
Hello! I hope this message finds you well. I have a question about adding additional metrics. I came across the documentation here: https://docs.vllm.ai/en/latest/serving/metrics.html, which introduces various metrics. I believe incorporating these could greatly help in diagnosing and addressing performance bottlenecks in the model. Would it be possible for you to provide some guidance or suggestions on how to add these metrics? I would sincerely appreciate your help. Thank you so much!